refactor(assert_plugin_utility): create package for assertPlugin utilities
