+++ app/src/test/kotlin/io/clappr/player/app/plugin/VideoInfoPluginTest.kt
@@ -6,8 +6,8 @@ import android.widget.TextView
 import io.clappr.player.BuildConfig
 import io.clappr.player.app.R
 import io.clappr.player.app.plugin.util.FakePlayback
-import io.clappr.player.app.plugin.util.assertHidden
-import io.clappr.player.app.plugin.util.assertShown
+import io.clappr.player.app.plugin.assertPlugin.assertUiPluginHidden
+import io.clappr.player.app.plugin.assertPlugin.assertUiPluginShown
 import io.clappr.player.base.BaseObject
 import io.clappr.player.base.Event
 import io.clappr.player.base.Options
@@ -101,28 +101,28 @@ class VideoInfoPluginTest {
         val newPlayback = FakePlayback(source)
         core.activeContainer?.playback = newPlayback
 
-        assertHidden(videoInfoPlugin)
+        assertUiPluginHidden(videoInfoPlugin)
     }
 
     @Test
     fun shouldShowViewWhenWillPlayEventIsTriggered() {
         core.activeContainer?.playback?.trigger(Event.WILL_PLAY.value)
 
-        assertShown(videoInfoPlugin)
+        assertUiPluginShown(videoInfoPlugin)
     }
 
     @Test
     fun shouldStopListeningOldPlaybackWhenDidChangePlaybackEventIsTriggered() {
         val oldPlayback = core.activePlayback
 
-        assertHidden(videoInfoPlugin)
+        assertUiPluginHidden(videoInfoPlugin)
 
         val newPlayback = FakePlayback(source)
         core.activeContainer?.playback = newPlayback
 
         oldPlayback?.trigger(Event.WILL_PLAY.value)
 
-        assertHidden(videoInfoPlugin)
+        assertUiPluginHidden(videoInfoPlugin)
     }
 
     @Test
@@ -133,6 +133,6 @@ class VideoInfoPluginTest {
 
         oldPlayback?.trigger(Event.WILL_PLAY.value)
 
-        assertHidden(videoInfoPlugin)
+        assertUiPluginHidden(videoInfoPlugin)
     }
 }
