+++ app/src/test/kotlin/io/clappr/player/app/plugin/assertPlugin/assertUiPlugin.kt
@@ -1,15 +1,15 @@
-package io.clappr.player.app.plugin.util
+package io.clappr.player.app.plugin.assertPlugin
 
 import android.view.View
 import io.clappr.player.plugin.UIPlugin
 import kotlin.test.assertEquals
 
-fun assertHidden(plugin: UIPlugin) {
+fun assertUiPluginHidden(plugin: UIPlugin) {
     assertEquals(UIPlugin.Visibility.HIDDEN, plugin.visibility)
     assertEquals(View.GONE, plugin.view?.visibility)
 }
 
-fun assertShown(plugin: UIPlugin) {
+fun assertUiPluginShown(plugin: UIPlugin) {
     assertEquals(UIPlugin.Visibility.VISIBLE, plugin.visibility)
     assertEquals(View.VISIBLE, plugin.view?.visibility)
 }
